ref: separate flow vs log tests
