+++ tests/trainer/test_train_loop_flow_dict_1_0.py
@@ -1,0 +1,186 @@
+"""
+Tests to ensure that the training loop works with a dict (1.0)
+"""
+from pytorch_lightning import Trainer
+from tests.base.deterministic_model import DeterministicModel
+import os
+import torch
+
+
+def test__training_step__flow_dict(tmpdir):
+    """
+    Tests that only training_step can be used
+    """
+    os.environ['PL_DEV_DEBUG'] = '1'
+
+    class TestModel(DeterministicModel):
+        def training_step(self, batch, batch_idx):
+            acc = self.step(batch, batch_idx)
+            acc = acc + batch_idx
+            self.training_step_called = True
+            return {'loss': acc, 'random_things': [1, 'a', torch.tensor(2)]}
+
+        def backward(self, trainer, loss, optimizer, optimizer_idx):
+            loss.backward()
+
+    model = TestModel()
+    model.val_dataloader = None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        limit_train_batches=2,
+        limit_val_batches=2,
+        max_epochs=2,
+        row_log_interval=1,
+        weights_summary=None,
+    )
+    trainer.fit(model)
+
+    # make sure correct steps were called
+    assert model.training_step_called
+    assert not model.training_step_end_called
+    assert not model.training_epoch_end_called
+
+
+def test__training_step__tr_step_end__flow_dict(tmpdir):
+    """
+    Tests that only training_step can be used
+    """
+    os.environ['PL_DEV_DEBUG'] = '1'
+
+    class TestModel(DeterministicModel):
+        def training_step(self, batch, batch_idx):
+            acc = self.step(batch, batch_idx)
+            acc = acc + batch_idx
+            self.training_step_called = True
+            self.out = {'loss': acc, 'random_things': [1, 'a', torch.tensor(2)]}
+            return self.out
+
+        def training_step_end(self, tr_step_output):
+            assert tr_step_output == self.out
+            assert self.count_num_graphs(tr_step_output) == 1
+            self.training_step_end_called = True
+            return tr_step_output
+
+        def backward(self, trainer, loss, optimizer, optimizer_idx):
+            loss.backward()
+
+    model = TestModel()
+    model.val_dataloader = None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        limit_train_batches=2,
+        limit_val_batches=2,
+        max_epochs=2,
+        row_log_interval=1,
+        weights_summary=None,
+    )
+    trainer.fit(model)
+
+    # make sure correct steps were called
+    assert model.training_step_called
+    assert model.training_step_end_called
+    assert not model.training_epoch_end_called
+
+
+def test__training_step__epoch_end__flow_dict(tmpdir):
+    """
+    Tests that only training_step can be used
+    """
+    os.environ['PL_DEV_DEBUG'] = '1'
+
+    class TestModel(DeterministicModel):
+        def training_step(self, batch, batch_idx):
+            acc = self.step(batch, batch_idx)
+            acc = acc + batch_idx
+
+            self.training_step_called = True
+            out = {'loss': acc, 'random_things': [1, 'a', torch.tensor(2)]}
+            return out
+
+        def training_epoch_end(self, outputs):
+            self.training_epoch_end_called = True
+
+            # verify we saw the current num of batches
+            assert len(outputs) == 2
+
+            for b in outputs:
+                assert isinstance(b, dict)
+                assert self.count_num_graphs(b) == 0
+                assert {'random_things', 'loss'} == set(b.keys())
+
+        def backward(self, trainer, loss, optimizer, optimizer_idx):
+            loss.backward()
+
+    model = TestModel()
+    model.val_dataloader = None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        limit_train_batches=2,
+        limit_val_batches=2,
+        max_epochs=2,
+        row_log_interval=1,
+        weights_summary=None,
+    )
+    trainer.fit(model)
+
+    # make sure correct steps were called
+    assert model.training_step_called
+    assert not model.training_step_end_called
+    assert model.training_epoch_end_called
+
+
+def test__training_step__step_end__epoch_end__flow_dict(tmpdir):
+    """
+    Tests that only training_step can be used
+    """
+    os.environ['PL_DEV_DEBUG'] = '1'
+
+    class TestModel(DeterministicModel):
+        def training_step(self, batch, batch_idx):
+            acc = self.step(batch, batch_idx)
+            acc = acc + batch_idx
+
+            self.training_step_called = True
+            self.out = {'loss': acc, 'random_things': [1, 'a', torch.tensor(2)]}
+            return self.out
+
+        def training_step_end(self, tr_step_output):
+            assert tr_step_output == self.out
+            assert self.count_num_graphs(tr_step_output) == 1
+            self.training_step_end_called = True
+            return tr_step_output
+
+        def training_epoch_end(self, outputs):
+            self.training_epoch_end_called = True
+
+            # verify we saw the current num of batches
+            assert len(outputs) == 2
+
+            for b in outputs:
+                assert isinstance(b, dict)
+                assert self.count_num_graphs(b) == 0
+                assert {'random_things', 'loss'} == set(b.keys())
+
+        def backward(self, trainer, loss, optimizer, optimizer_idx):
+            loss.backward()
+
+    model = TestModel()
+    model.val_dataloader = None
+
+    trainer = Trainer(
+        default_root_dir=tmpdir,
+        limit_train_batches=2,
+        limit_val_batches=2,
+        max_epochs=2,
+        row_log_interval=1,
+        weights_summary=None,
+    )
+    trainer.fit(model)
+
+    # make sure correct steps were called
+    assert model.training_step_called
+    assert model.training_step_end_called
+    assert model.training_epoch_end_called
